Move a few llm helpers out of stdlib
