+++ catgrad/src/stdlib/nn/mod.rs
@@ -244,84 +244,3 @@ pub fn linear_no_bias(builder: &Builder, in_dim: usize, out_dim: usize, p: Path,
 pub fn linear(builder: &Builder, in_dim: usize, out_dim: usize, p: Path, x: Var) -> Var {
     linear_b(builder, in_dim, out_dim, true, p, x)
 }
-
-pub fn layernorm_raw(builder: &Builder, eps: f32, x: Var) -> Var {
-    let x_shape = shape(builder, x.clone());
-    let [_, _, n] = unpack::<3>(builder, x_shape.clone());
-    let s = sum(builder, x.clone());
-
-    let constn = nat_to_u32(builder, n);
-    let constn = cast(builder, constn, dtype(builder, x.clone()));
-    let sh = shape(builder, s.clone());
-    let constn = broadcast(builder, constn, sh);
-
-    let mean = s / constn.clone();
-    let nom = x - broadcast(builder, mean, x_shape.clone());
-
-    let var = sum(builder, nom.clone() * nom.clone()) / constn;
-    let sh = shape(builder, var.clone());
-    let epsilon = constant(builder, eps, &sh);
-    let stddev = sqrt(builder, var + epsilon);
-    let denom = broadcast(builder, stddev, x_shape);
-
-    nom / denom
-}
-
-pub fn layernorm(builder: &Builder, eps: f32, p: Path, x: Var) -> Var {
-    let gamma = param(builder, &p.extend(["weight"]).unwrap());
-    let lr = layernorm_raw(builder, eps, x);
-    let lr_shape = shape(builder, lr.clone());
-    let gamma = broadcast(builder, gamma, lr_shape.clone());
-    let lr = lr * gamma;
-
-    let beta = param(builder, &p.extend(["bias"]).unwrap());
-    let beta = broadcast(builder, beta, lr_shape);
-    lr + beta
-}
-
-pub fn rmsnorm_raw(builder: &Builder, eps: f32, x: Var) -> Var {
-    let x_shape = shape(builder, x.clone());
-    let [_, _, n] = unpack::<3>(builder, x_shape.clone());
-    let s = sum(builder, x.clone() * x.clone());
-
-    let constn = nat_to_u32(builder, n);
-    let constn = cast(builder, constn, dtype(builder, x.clone()));
-    let sh = shape(builder, s.clone());
-    let constn = broadcast(builder, constn, sh);
-
-    let mean = s / constn;
-
-    let epsilon = constant(builder, eps, &shape(builder, mean.clone()));
-    let rms = sqrt(builder, mean + epsilon);
-    let denom = broadcast(builder, rms, x_shape);
-    x / denom
-}
-
-// rmsnorm(x) = x / √(E[x²] + ε) × γ
-pub fn rmsnorm(builder: &Builder, eps: f32, p: Path, x: Var) -> Var {
-    let gamma = param(builder, &p.extend(["weight"]).unwrap());
-    let lr = rmsnorm_raw(builder, eps, x);
-    let lr_shape = shape(builder, lr.clone());
-    let gamma = broadcast(builder, gamma, lr_shape);
-    lr * gamma
-}
-
-/// Remove a dimension of extent 1 to a tensor
-pub fn squeeze<const N: usize, const M: usize>(builder: &Builder, dim: usize, x: Var) -> Var {
-    assert_eq!(N, M + 1);
-    let x_shape = shape(builder, x.clone());
-    let mut s = unpack::<N>(builder, x_shape).to_vec();
-    s.remove(dim);
-    let new_shape = pack::<M>(builder, s.try_into().unwrap());
-    reshape(builder, new_shape, x)
-}
-
-/// Add an additional dimension of extent 1 to a tensor
-pub fn unsqueeze<const N: usize, const M: usize>(builder: &Builder, dim: usize, x: Var) -> Var {
-    assert_eq!(N + 1, M);
-    let x_shape = shape(builder, x.clone());
-    let mut s = unpack::<N>(builder, x_shape).to_vec();
-    s.insert(dim, 1.to_nat(builder));
-    let new_shape = pack::<M>(builder, s.try_into().unwrap());
-    reshape(builder, new_shape, x)
-}
